+++ src/extension.ts
@@ -21,7 +21,7 @@ export function activate(p_context: vscode.ExtensionContext) {
     const w_leoIntegExtension = vscode.extensions.getExtension(Constants.PUBLISHER + '.' + Constants.NAME)!;
     const w_leoIntegVersion = w_leoIntegExtension.packageJSON.version;
     const w_leo: LeoIntegration = new LeoIntegration(p_context);
-    const w_leoSettingsWebview: LeoSettingsProvider = new LeoSettingsProvider(p_context, w_leo);
+    const w_leoSettingsWebview: LeoSettingsProvider = w_leo.leoSettingsWebview;
     const w_previousVersion = p_context.globalState.get<string>(Constants.VERSION_STATE_KEY);
     const w_start = process.hrtime(); // For calculating total startup time duration
 
@@ -44,7 +44,6 @@ export function activate(p_context: vscode.ExtensionContext) {
         body: true,
         states: true
     };
-    const showInfo = vscode.window.showInformationMessage;
 
     const w_commands: [string, (...args: any[]) => any][] = [