+++ src/webviews/settingsPanel/index.ts
@@ -105,15 +105,16 @@ declare function acquireVsCodeApi(): VsCodeApi;
             return onInputChanged(this);
         });
         listenAll('select[data-setting]', 'change', function (this: HTMLSelectElement) {
-            return onInputSelected(this);
+            return onDropdownChanged(this);
         });
     }
 
-    function onInputSelected(element: HTMLSelectElement) {
+    function onDropdownChanged(element: HTMLSelectElement) {
         if (element) {
             const w_value = element.options[element.selectedIndex].value;
             frontConfig[element.id] = w_value;
         }
+        dirty!.className = "show";
         applyChanges();
     }
 
@@ -215,8 +216,6 @@ declare function acquireVsCodeApi(): VsCodeApi;
     function getSettingValue(p_setting: string): any {
         return frontConfig[p_setting];
     }
-
-
 
     var applyChanges = debounce(function () {
         var w_changes: ConfigSetting[] = [];